djangoresttest/requirements.txt
django
djangorestframework
environs
python-dotenv